requirements.txt
attrs==21.2.0
aws-cdk.assets==1.126.0
aws-cdk.aws-apigateway==1.126.0
aws-cdk.aws-applicationautoscaling==1.126.0
aws-cdk.aws-autoscaling-common==1.126.0
aws-cdk.aws-certificatemanager==1.126.0
aws-cdk.aws-cloudformation==1.126.0
aws-cdk.aws-cloudwatch==1.126.0
aws-cdk.aws-codeguruprofiler==1.126.0
aws-cdk.aws-codestarnotifications==1.126.0
aws-cdk.aws-cognito==1.126.0
aws-cdk.aws-ec2==1.126.0
aws-cdk.aws-ecr==1.126.0
aws-cdk.aws-ecr-assets==1.126.0
aws-cdk.aws-efs==1.126.0
aws-cdk.aws-elasticloadbalancingv2==1.126.0
aws-cdk.aws-events==1.126.0
aws-cdk.aws-iam==1.126.0
aws-cdk.aws-kms==1.126.0
aws-cdk.aws-lambda==1.126.0
aws-cdk.aws-logs==1.126.0
aws-cdk.aws-route53==1.126.0
aws-cdk.aws-s3==1.126.0
aws-cdk.aws-s3-assets==1.126.0
aws-cdk.aws-signer==1.126.0
aws-cdk.aws-sns==1.126.0
aws-cdk.aws-sqs==1.126.0
aws-cdk.aws-ssm==1.126.0
aws-cdk.cloud-assembly-schema==1.126.0
aws-cdk.core==1.126.0
aws-cdk.custom-resources==1.126.0
aws-cdk.cx-api==1.126.0
aws-cdk.region-info==1.126.0
cattrs==1.8.0
constructs==3.3.160
jsii==1.36.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.2
